Applied alpha
Now the image will consider the alpha value inversely when averaging the luminance. This removes the artifacting in the borders of the image, but leads to weaker border values, I had to add code which extends the original image at its border by copying the pixels, to account for this.