Fix: Fixed the previous fix
After changing the line width from int to double it lead to situations where the line would have a non zero width that was smaller than 1 pixel. This made a bunch of ugly gray lines in areas which should be white. Fixed by adding a threshold value to the width of the line.